@@ -18,8 +18,10 @@
     "build": "rimraf dist && npm-run-all --parallel --silent build:ts build:editor",
     "build:ts": "tsc",
     "build:editor": "webpack --config src/ui/editor/webpack.config.js",
+    "build:editor:wds": "webpack-dev-server --config src/ui/editor/webpack.config.js --port 9000",
     "app": "tsc && npm run -s build:editor && electron bin/electron",
-    "app:debug": "npm run -s app -- --debug",
+    "app:debug:only": "electron bin/electron --debug",
+    "app:debug": "npm-run-all --parallel --silent build:editor:wds app:debug:only",
     "prepare": "npm run -s build"
   },
   "nyc": {
@@ -56,18 +58,17 @@
   },
   "homepage": "https://github.com/mixer/miix-cli#readme",
   "devDependencies": {
-    "@angular/animations": "^4.3.4",
-    "@angular/cdk": "^2.0.0-beta.12",
-    "@angular/common": "^4.3.4",
-    "@angular/compiler": "^4.3.4",
-    "@angular/core": "^4.3.4",
-    "@angular/forms": "^4.3.4",
-    "@angular/http": "^4.3.4",
-    "@angular/material": "^2.0.0-beta.12",
-    "@angular/platform-browser": "^4.3.4",
-    "@angular/platform-browser-dynamic": "^4.3.4",
-    "@ngrx/effects": "^4.1.1",
-    "@ngrx/store": "^4.0.2",
+    "@angular/animations": "^5.2.8",
+    "@angular/cdk": "^5.2.4",
+    "@angular/common": "^5.2.8",
+    "@angular/compiler": "^5.2.8",
+    "@angular/core": "^5.2.8",
+    "@angular/forms": "^5.2.8",
+    "@angular/material": "^5.2.4",
+    "@angular/platform-browser": "^5.2.8",
+    "@angular/platform-browser-dynamic": "^5.2.8",
+    "@ngrx/effects": "^5.2.0",
+    "@ngrx/store": "^5.2.0",
     "@types/chai": "^4.0.3",
     "@types/chai-as-promised": "0.0.31",
     "@types/chai-subset": "^1.3.0",
@@ -93,7 +94,7 @@
     "@types/tar": "^4.0.0",
     "@types/yargs": "^8.0.2",
     "angular2-template-loader": "^0.6.2",
-    "awesome-typescript-loader": "^3.2.2",
+    "awesome-typescript-loader": "^5.0.0-1",
     "bluefill": "^0.1.3",
     "chai": "^4.1.1",
     "chai-as-promised": "^7.1.1",
@@ -105,6 +106,7 @@
     "eventemitter3": "^2.0.3",
     "fast-json-patch": "^2.0.4",
     "highlight.js": "^9.12.0",
+    "html-webpack-plugin": "^3.0.6",
     "mocha": "^3.5.0",
     "moment": "^2.18.1",
     "node-sass": "^4.5.3",
@@ -113,7 +115,7 @@
     "nyc": "^11.2.1",
     "prettier": "^1.5.3",
     "raw-loader": "^0.5.1",
-    "rxjs": "^5.4.3",
+    "rxjs": "^5.5.6",
     "sass-loader": "^6.0.6",
     "screenfull": "^3.3.1",
     "sinon": "^3.2.0",
@@ -124,7 +126,9 @@
     "tslint": "^5.6.0",
     "tslint-microsoft-contrib": "5.0.1",
     "url-loader": "^0.5.9",
-    "webpack": "^3.5.4",
+    "webpack": "^4.1.1",
+    "webpack-cli": "^2.0.11",
+    "webpack-dev-server": "^3.1.1",
     "zone.js": "^0.8.16"
   },
   "dependencies": {
